+++ semicolon/src/main/java/dukku/semicolon/shared/payment/dto/PaymentResponse.java
@@ -64,7 +64,7 @@ public class PaymentResponse {
         /**
          * PG 결제 금액
          */
-        private Integer amount;
+        private Long amount;
 
         /**
          * 주문명
@@ -94,16 +94,16 @@ public class PaymentResponse {
         /**
          * 최종 결제 금액
          */
-        private Integer finalPayAmount;
+        private Long finalPayAmount;
 
         /**
          * 예치금 사용 금액
          */
-        private Integer depositUseAmount;
+        private Long depositUseAmount;
 
         /**
          * PG 결제 금액
          */
-        private Integer pgPayAmount;
+        private Long pgPayAmount;
     }
 }
